Add custom message factory support
Use this to change the DefaultGelfMessageFactory behaviour.

Example:

    <appender name="graylogMessages" class="com.github.pukkaone.gelf.logback.GelfAppender">
      ...
      <marshaller class="com.example.MyGelfMessageFactory">
        <mdcTimestampProperty>ReceivedDateTime</mdcTimestampProperty>
      </marshaller>
      ...
    </appender>
